requirements.txt
pandas
numpy
pyarrow
scikit-learn
mlflow
evidently
confluent-kafka
websockets
python-dotenv
